Fixed ressource name format when passed to prompt Fixed typo
