+++ overlore/eternum/constants.py
@@ -1,21 +1,4 @@
-from __future__ import annotations
-
-import json
-import logging
-from enum import Enum
-from typing import TypedDict, cast
-
-from overlore.eternum.types import RealmPosition
-
 DAY_IN_SECONDS: int = 24 * 60 * 60
-
-
-class RealmGeodata(TypedDict):
-    xy: list[float]
-    name: str
-
-
-logger = logging.getLogger("overlore")
 
 ORDERS = [
     "power",
@@ -37,77 +20,30 @@ ORDERS = [
 ]
 
 
-class ResourceIds(Enum):
-    Wood = 1
-    Stone = 2
-    Coal = 3
-    Copper = 4
-    Obsidian = 5
-    Silver = 6
-    Ironwood = 7
-    ColdIron = 8
-    Gold = 9
-    Hartwood = 10
-    Diamonds = 11
-    Sapphire = 12
-    Ruby = 13
-    DeepCrystal = 14
-    Ignium = 15
-    EtherealSilica = 16
-    TrueIce = 17
-    TwilightQuartz = 18
-    AlchemicalSilver = 19
-    Adamantine = 20
-    Mithral = 21
-    Dragonhide = 22
-    Lords = 253
-    Wheat = 254
-    Fish = 255
-
-
-class Winner(Enum):
-    Attacker = 0
-    Target = 1
-
-
-class Realms:
-    _instance = None
-    geodata: list[RealmGeodata]
-
-    @classmethod
-    def instance(cls) -> Realms:
-        if cls._instance is None:
-            logger.debug("Creating Realms constants class")
-            cls._instance = cls.__new__(cls)
-        return cls._instance
-
-    def __init__(self) -> None:
-        raise RuntimeError("Call instance() instead")
-
-    def load_geodata(self, path: str) -> list[RealmGeodata]:
-        with open(path) as file:
-            geodata_file = json.load(file)
-            features: list[RealmGeodata] = geodata_file["features"]
-            return features
-
-    def init(self, path: str = "./data/realms_geodata.json") -> Realms:
-        self.geodata = self.load_geodata(path)
-        return self
-
-    def order_by_order_id(self, i: int) -> str:
-        return ORDERS[i - 1]
-
-    def name_by_id(self, i: int) -> str:
-        if i <= 0 or i > 8000 or self.geodata[i - 1].get("xy") is None:
-            raise RuntimeError("Error with input or geodata")
-        return str(self.geodata[i - 1]["name"])
-
-    def position_by_id(self, i: int) -> RealmPosition:
-        if i <= 0 or i > 8000 or self.geodata[i - 1].get("xy") is None:
-            raise RuntimeError("Error with input or geodata")
-        xy: list[float] = cast(list[float], self.geodata[i - 1]["xy"])
-        ret: tuple[float, float] = (
-            float(xy[0]),
-            float(xy[1]),
-        )
-        return ret
+ResourceNamesById = {
+    1: "wood",
+    2: "stone",
+    3: "coal",
+    4: "copper",
+    5: "obsidian",
+    6: "silver",
+    7: "ironwood",
+    8: "cold iron",
+    9: "gold",
+    10: "hartwood",
+    11: "diamonds",
+    12: "sapphire",
+    13: "ruby",
+    14: "deep crystal",
+    15: "ignium",
+    16: "ethereal silica",
+    17: "true ice",
+    18: "twilight quartz",
+    19: "alchemical silver",
+    20: "adamantine",
+    21: "mithral",
+    22: "dragonhide",
+    253: "Lords",
+    254: "wheat",
+    255: "fish",
+}
