+++ overlore/eternum/types.py
@@ -1,3 +1,4 @@
+from enum import Enum
 from typing import TypeAlias, TypedDict
 
 RealmPosition: TypeAlias = tuple[float, float]
@@ -12,3 +13,8 @@ class ResourceAmount(TypedDict):
 
 
 ResourceAmounts: TypeAlias = list[ResourceAmount]
+
+
+class Winner(Enum):
+    Attacker = 0
+    Target = 1
