+++ overlore/utils.py
@@ -5,6 +5,7 @@ from starknet_py.cairo.felt import encode_shortstring
 from starknet_py.hash.utils import ECSignature, message_signature
 from starknet_py.hash.utils import compute_hash_on_elements as pedersen
 
+from overlore.eternum.constants import ResourceNamesById
 from overlore.types import Characteristics
 
 logger = logging.getLogger("overlore")
@@ -13,11 +14,11 @@ U2_MASK: int = 0x3
 U8_MASK: int = 0xFF
 
 
-def get_enum_name_by_value(enum: Any, val: Any) -> str:
-    for enum_entry in enum:
-        if enum_entry.value == val:
-            return str(enum_entry.name)
-    raise RuntimeError(f"No value for val {val} in enum")
+def get_ressource_name_by_id(resource_id: Any) -> str:
+    if resource_id in ResourceNamesById:
+        return ResourceNamesById[resource_id]
+    else:
+        raise RuntimeError(f"No resource found for ID {resource_id}")
 
 
 def sign_parameters(msg: Sequence, priv_key: str) -> ECSignature:
